Address issue with different parts that share same Uri.
There is an obscure bug that occurs when a part is encountered by the GenerateSourceCode method that has already been encountered earlier in the process.  The part in question is separated by 1 or more parts further down in the document.  While here, add more methods and fields to aid in the comparison process.